Zbfill: finished and cleaned up
The code now works and fills all holes as intended. The code was tested
on drives of varying sizes (256MiB, 512MiB, 1GiB and 2GiB) and appears
to work as intended. Drives of both 1024-sized and 4096-sized blocks
were tested. Also, drives with sparse_super_2 and bigalloc features were
briefly tested. TODO: test drive sizes that are not powers of 2. Also,
json output was added.